chore: refactored and rehomed media sorting functionality and components
